CI refactorr
